Moved App.js and App.test.js to src directory
