+++ README.md
@@ -21,7 +21,7 @@ built at a later date.
 
 **NOTE**: We will use a mock api instead of a real rest api. This will allow us some nice flexibility throughout the tutorial.
 
-#### Writing our first unit test
+#### Setting up our unit testing framework
 
 I prefer to use enzyme to test react components as I find the api for enzyme to be more natural than the one provided with react-test-renderer. Let's go ahead and add enzyme to our packages:
 
@@ -50,5 +50,10 @@ import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 configure({ adapter: new Adapter() });
 ```
+Okay now our framework is all setup and we can begin writing tests against our application. 
+
+#### Writing our first test
+
+My preference is to have a [src](./src) directory for all source code. Let's move our [App.js](src/App.js) and [App.test.js](src/App.test.js) to our [src](./src) directory.
